feature(zod): moved "lite validation" into new folder
