+++ packages/groq-builder/src/validation/lite/array-shape.test.ts
