+++ packages/groq-builder/src/validation/lite/index.ts
@@ -3,8 +3,6 @@ import { sanityValidation } from "./sanity-content-blocks";
 import { objectValidation } from "./object-shape";
 import { arrayValidation } from "./array-shape";
 
-export { zod } from "./zod";
-
 export const validation = {
   ...primitiveValidation,
   ...sanityValidation,
